+++ src/main/java/dev/root101/clean/core/utils/validation/ValidationService.java
@@ -1,12 +1,10 @@
 package dev.root101.clean.core.utils.validation;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.root101.clean.core.exceptions.ValidationException;
 import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 import javax.validation.Configuration;
 import javax.validation.ConstraintViolation;
 import javax.validation.MessageInterpolator;
@@ -15,6 +13,14 @@ import javax.validation.Validator;
 import org.springframework.util.ClassUtils;
 
 public class ValidationService {
+
+    public static final String PARENT_TREE_SEPARATOR = ".";
+
+    public record TracedViolation(
+            Set<ConstraintViolation<Object>> currentViolations,
+            String parentTree) {
+
+    }
 
     //store the config
     private static final Configuration<?> CONFIG = Validation.byDefaultProvider().configure();
@@ -32,52 +38,101 @@ public class ValidationService {
         ).buildValidatorFactory().getValidator();
     }
 
-    public static List<ConstraintViolation<Object>> validate(Object... objects) {
-        List<ConstraintViolation<Object>> errors = new ArrayList<>();
+    public static void validateAndThrow(Object... objects) throws ValidationException {
+        List<TracedViolation> errors = validate(objects);
+        if (!errors.isEmpty()) {
+            throw new ValidationException(convertMessages(errors));
+        }
+    }
 
-        for (Object object : objects) {
-            Set<ConstraintViolation<Object>> aaa = DEFAULT_VALIDATOR.validate(object);
-            errors.addAll(aaa);
+    public static List<TracedViolation> validate(Object... objects) {
+        List<TracedViolation> errors = new ArrayList<>();
+
+        if (objects.length == 1) {
+            errors.add(
+                    new TracedViolation(
+                            DEFAULT_VALIDATOR.validate(objects[0]),
+                            ""
+                    )
+            );
+        } else {
+            for (int i = 0; i < objects.length; i++) {
+                errors.add(
+                        new TracedViolation(
+                                DEFAULT_VALIDATOR.validate(objects[i]),
+                                "[%s]".formatted(i)
+                        )
+                );
+            }
         }
 
         return errors;
     }
 
-    public static List<ConstraintViolation<Object>> validateRecursive(Object... objects) {
-        return validateRecursive(new ArrayList<>(), objects);
-    }
-
-    public static void validateAndThrow(Object... objects) throws ValidationException {
-        List<ConstraintViolation<Object>> errors = validate(objects);
-        if (!errors.isEmpty()) {
-            throw new ValidationException(convertMessages(errors));
-        }
-    }
-
     public static void validateRecursiveAndThrow(Object... objects) throws ValidationException {
-        List<ConstraintViolation<Object>> errors = validateRecursive(objects);
+        List<TracedViolation> errors = validateRecursive(objects);
         if (!errors.isEmpty()) {
             throw new ValidationException(convertMessages(errors));
         }
     }
 
-    public static List<ValidationException.ValidationErrorMessage> convertMessages(List<ConstraintViolation<Object>> violation) {
-        return violation.stream().map((viol) -> {
-            String fieldName = viol.getPropertyPath().toString();
-            try {
-                Field field = viol.getRootBeanClass().getDeclaredField(fieldName);
-                ValidationFieldName fieldNameAnnotation = field.getDeclaredAnnotation(ValidationFieldName.class);
-                if (fieldNameAnnotation != null) {
-                    fieldName = fieldNameAnnotation.value();
-                }
-            } catch (NoSuchFieldException | SecurityException e) {
-                System.out.println("Error convirtiendo los mensajes de las validaciones. " + e.getMessage());
+    public static List<TracedViolation> validateRecursive(Object... objects) {
+        if (objects.length == 1) {
+            return validateRecursive(new ArrayList<>(), "", objects[0]);
+        } else {
+            List<TracedViolation> errors = new ArrayList<>();
+            for (int i = 0; i < objects.length; i++) {
+                errors.addAll(
+                        validateRecursive(
+                                new ArrayList<>(),
+                                "[%s]".formatted(i),
+                                objects[i]
+                        )
+                );
             }
-            return new ValidationException.ValidationErrorMessage(fieldName, viol.getInvalidValue() == null ? "null" : viol.getInvalidValue().toString(), viol.getMessage());
-        }).collect(Collectors.toList());
+            return errors;
+        }
     }
 
-    private static List<ConstraintViolation<Object>> validateRecursive(List<ConstraintViolation<Object>> currentViolations, Object... objects) {
+    public static List<ValidationException.ValidationErrorMessage> convertMessages(List<TracedViolation> violation) {
+        List<ValidationException.ValidationErrorMessage> messages = new ArrayList<>();
+        for (TracedViolation general : violation) {
+            for (ConstraintViolation<Object> viol : general.currentViolations) {
+
+                String fieldName = viol.getPropertyPath().toString();
+                try {
+                    Field field = viol.getRootBeanClass().getDeclaredField(fieldName);
+                    ValidationFieldName fieldNameAnnotation = field.getDeclaredAnnotation(ValidationFieldName.class);
+                    if (fieldNameAnnotation != null) {
+                        fieldName = fieldNameAnnotation.value();
+                    }
+                } catch (NoSuchFieldException | SecurityException e) {
+                    System.out.println("Error convirtiendo los mensajes de las validaciones. " + e.getMessage());
+                }
+
+                String source = general.parentTree + PARENT_TREE_SEPARATOR + fieldName;
+                if (source.startsWith(PARENT_TREE_SEPARATOR)) {
+                    source = source.substring(1, source.length());
+                } else if (source.startsWith("[")) {
+                    source = "root" + source;
+                }
+
+                messages.add(
+                        new ValidationException.ValidationErrorMessage(
+                                source,
+                                viol.getInvalidValue() == null
+                                ? "null"
+                                : viol.getInvalidValue().toString(),
+                                viol.getMessage()
+                        )
+                );
+
+            }
+        }
+        return messages;
+    }
+
+    private static List<TracedViolation> validateRecursive(List<TracedViolation> violations, String parentTree, Object... objects) {
         //recorro la lista de objetos a validar
         for (Object object : objects) {
             //si NO es null lo proceso. Null no tiene sentido validarlo(DEFAULT_VALIDATOR.validate(null) lanza excepcion), se deberia haber validado una capa arriba.
@@ -85,38 +140,55 @@ public class ValidationService {
                 //si no es null compruebo si:
                 if (object instanceof Object[] arr) {
                     //es una instancia de arreglo, llamo a la recursividad con el arreglo
-                    validateRecursive(currentViolations, arr);
+                    for (int i = 0; i < arr.length; i++) {
+                        validateRecursive(violations, parentTree + "[%s]".formatted(i), arr);
+                    }
                 } else if (object instanceof List list) {
                     //es una instancia de lista, la convierto en arreglo y llamo a la recursividad con el arreglo
-                    validateRecursive(currentViolations, list.toArray());
+                    for (int i = 0; i < list.size(); i++) {
+                        validateRecursive(violations, parentTree + "[%s]".formatted(i), list.toArray());
+                    }
                 } else {
                     //no es ni una lista ni un arreglo, valido el objeto como objeto
-                    currentViolations.addAll(DEFAULT_VALIDATOR.validate(object));
+                    Set<ConstraintViolation<Object>> validations = DEFAULT_VALIDATOR.validate(object);
+                    if (!validations.isEmpty()) {
+                        violations.add(
+                                new TracedViolation(
+                                        validations,
+                                        parentTree
+                                )
+                        );
+                    }
 
                     //luego recorro todos sus campos a ver si alguno no es primitivo
                     Field fields[] = object.getClass().getDeclaredFields();
-                    Object[] arr
-                            = List.of(fields).stream()
-                                    //convierto los field en el objeto como tal
-                                    .map((field) -> {
-                                        field.setAccessible(true);
-                                        try {
-                                            return field.get(object);
-                                        } catch (IllegalAccessException | IllegalArgumentException e) {
-                                            System.out.println("Nunca debe entrar aqui");
-                                        }
-                                        return null;
-                                    })
-                                    //filtro de todos los valores los que NO sean primitivos
-                                    .filter((t) -> {
-                                        return t != null && !(ClassUtils.isPrimitiveOrWrapper(t.getClass()) || t.getClass().getName().startsWith("java"));
-                                    })
-                                    .toArray();
-                    //de todos los campos que no son promitivos los valido
-                    validateRecursive(currentViolations, arr);
+                    for (Field field : fields) {
+                        try {
+                            field.setAccessible(true);
+                            Object t = field.get(object);
+
+                            if (t != null
+                                    && (!(ClassUtils.isPrimitiveOrWrapper(t.getClass()) || t.getClass().getName().startsWith("java"))
+                                    || List.class.isAssignableFrom(t.getClass()))) {
+
+                                String fieldName = field.getName();
+
+                                ValidationFieldName fieldNameAnnotation = field.getDeclaredAnnotation(ValidationFieldName.class);
+                                if (fieldNameAnnotation != null) {
+                                    fieldName = fieldNameAnnotation.value();
+                                }
+
+                                //de todos los campos que no son primitivos los valido
+                                validateRecursive(violations, parentTree + PARENT_TREE_SEPARATOR + fieldName, t);
+                            }
+                        } catch (Exception e) {
+                            System.out.println("Nunca debe entrar aqui, si entra IGNORAR, es un problema de acceso a los fields del objeto. " + e.getMessage());
+                        }
+
+                    }
                 }
             }
         }
-        return currentViolations;
+        return violations;
     }
 }