+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+* 4.6.1.RELEASE.202220125:
+    * **REPO** :
+        * **IMPROVEMENT** :raised: : `findById` with id=null return null. Don't throw exception.
+    * **VALIDATIONS** :
+        * **IMPROVEMENT** :raised: : ValidationService mantain records of current tree in order to show complete source of errors fields.
+
 * 4.6.0.SNAPSHOT.202220113:
     * **GENERAL** :
         * **REFACTOR** :building: : Removed all standars for REST & UC layer.